refactor(progress): component-plane sweep with permission-key and goal-status fixes
