+++ packages/extension/src/progressView/frontend/components/streamHeaderView.ts
@@ -2,9 +2,9 @@
  * Shared `<stream-header>` wiring for the stream-content containers.
  *
  * The tool-use, workflow, and process containers each render an identical
- * eight-property `<stream-header>`; the only variation is the bypass/goal
- * flags, which exist only on tool-use state. Centralizing it here keeps the
- * header contract in one place so adding a prop touches a single call site.
+ * `<stream-header>`; the only variation is the bypass/goal fields, which
+ * exist only on tool-use state. Centralizing it here keeps the header
+ * contract in one place so adding a prop touches a single call site.
  */
 
 // Third-party imports
@@ -39,6 +39,8 @@ export function renderStreamHeader(
       .yoloActive=${Boolean(toolUse?.toolEditBypass)}
       .superYoloActive=${Boolean(toolUse?.superYoloBypass)}
       .goalActive=${Boolean(toolUse?.goalActive)}
+      .goalStatus=${toolUse?.goalStatus ?? ''}
+      .goalObjective=${toolUse?.goalObjective ?? ''}
       .unsupportedCommands=${unsupportedCommands}
     ></stream-header>
   `;
